added cohortowen subclass
Added cohortowen subclass of cohortshapley.  It takes a union_structure in addition to what the superclass needs.  The union structure must be a list of list of integers.  For an example of it working, see uploaded notebook cohortowen.

Known issue: the "permutations" warning comes during compute so will always be flagged even if doing the full cohortowen without provided permutations.